refactor/#28: 폴더 구조 수정
